Share argparse validators
